Update changelog with new features and changes; refactor JournalEntry component for editing functionality; enhance routing in App component; consolidate CSS styles and remove unused files; update package dependencies for compatibility.

+++ src/components/EntryCard.jsx
@@ -6,9 +6,9 @@ import '../css/EntryCard.css'
 const EntryCard = ({ entry }) => {
   const formatDate = (dateString) => {
     return new Date(dateString).toLocaleDateString('en-US', {
-      weekday: 'long',
+      weekday: 'short',
       year: 'numeric',
-      month: 'long',
+      month: 'numeric',
       day: 'numeric',
     })
   }
